fixed overflow errors; now getting same result as ECTA code

+++ ACESimBase/GameSolvingSupport/ECTAAlgorithm/MultiprecisionStatic.cs
@@ -81,7 +81,7 @@ namespace ACESimBase.GameSolvingSupport.ECTAAlgorithm
 			try
 			{
 				result = 0;
-				int placeValue = 1;
+				long placeValue = 1;
 				int digits = length(a);
 				for (int i = 1; i <= digits - 1; i++)
 				{
